Use flask_app from conftest
